+++ packages/geometry-engine/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @mlightcad/geometry-engine
 
+## 3.0.9
+
+### Patch Changes
+
+- Updated dependencies
+  - @mlightcad/common@1.2.7
+
 ## 3.0.8
 
 ### Patch Changes